+++ NES Simulator/Memory.h
@@ -13,7 +13,7 @@ private:
 	uint8_t* ram = NULL;
 
 public:
-	uint8_t Read(uint32_t address) const;
-	void Write(uint32_t address, uint8_t data);
+	uint8_t Read(uint16_t address) const;
+	void Write(uint16_t address, uint8_t data);
 };
